Change selectableviews package to views-selectable
This commit also rename the modules
