+++ views-selectable-appcompat/src/main/AndroidManifest.xml
@@ -14,4 +14,4 @@
   ~ limitations under the License.
   -->
 
-<manifest package="splitties.selectableviews.appcompat"/>
+<manifest package="splitties.views.selectable.appcompat"/>
