+++ views-selectable/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <!--
-  ~ Copyright (c) 2018. Louis Cognault Ayeva Derman
+  ~ Copyright (c) 2016. Louis Cognault Ayeva Derman
   ~
   ~ Licensed under the Apache License, Version 2.0 (the "License");
   ~ you may not use this file except in compliance with the License.
@@ -14,4 +14,4 @@
   ~ limitations under the License.
   -->
 
-<manifest package="splitties.selectableviews.constraintlayout"/>
+<manifest package="splitties.views.selectable"/>
