+++ views-selectable/src/main/java/splitties/views/selectable/SelectableTextView.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018. Louis Cognault Ayeva Derman
+ * Copyright (c) 2017. Louis Cognault Ayeva Derman
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package splitties.selectableviews.appcompat
+package splitties.views.selectable
 
 import android.content.Context
 import android.graphics.Canvas
@@ -22,13 +22,16 @@ import android.graphics.drawable.Drawable
 import android.os.Build.VERSION_CODES.LOLLIPOP
 import android.support.annotation.CallSuper
 import android.support.annotation.RequiresApi
-import android.support.v7.widget.AppCompatTextView
 import android.util.AttributeSet
+import android.widget.TextView
 import splitties.resources.styledDrawable
 
+/**
+ * [TextView] with ripple effect / select foreground when touched.
+ */
 open class SelectableTextView @JvmOverloads constructor(
         context: Context, attrs: AttributeSet?, defStyleAttr: Int = 0
-) : AppCompatTextView(context, attrs, defStyleAttr) {
+) : TextView(context, attrs, defStyleAttr) {
     constructor(context: Context) : this(context, null)
 
     var foregroundSelector: Drawable? = null
